Corrected linter rules
